TechBook/SkillLink_Examples/requirements.txt
python-dotenv
openai
google-genai
requests
SkillLink
BitSig
